Move DataPlaneIcon into shared location
